refactor: schemas をゲームごとのファイルに分離
`schemas/gameData.ts` に混在していた Game1/2/3 の zod スキーマを
`schemas/games/<game>.ts` の 3 ファイルに分離する。`gameData.ts` は
後方互換のための barrel として再エクスポートのみ残し、既存 import
パスを壊さない。

- termsGame.ts: ScrollEvent / CheckboxState / PopupStats / Game1Data
- helpdeskGame.ts: Game2InputMethod / Game2Turn / Game2TextInputMetrics / Game2Data
- groupChatGame.ts: Game3Stage / Game3Data

OpenAPI 出力に影響しないことを `npm run check:api-types` で確認済み
（registry へのコンポーネント名・登録順を維持）。
